Create html code of second page
This is the html code of second page. The user will enter this page after clicking the comment button on the first page. Here the title and description of his problem will be displayed .He himself and other members can comment on his problem by entering their username and email.